Interpreter pattern was implemented
