+++ src/main/java/com/example/shell/fx/layouts/files/filters/Expression.java
@@ -1,0 +1,5 @@
+package com.example.shell.fx.layouts.files.filters;
+
+public abstract class Expression<T> {
+    public abstract boolean interpret(T context);
+}
