+++ src/main/java/com/example/shell/fx/layouts/files/filters/FileFilterExpression.java
@@ -1,0 +1,7 @@
+package com.example.shell.fx.layouts.files.filters;
+
+import java.io.File;
+
+public abstract class FileFilterExpression extends Expression<File> {
+    public abstract boolean interpret(File file);
+}
