--- src/main/java/com/example/shell/fx/layouts/files/view/comboBoxes/ComboBoxTemplate.java
@@ -1,20 +1,0 @@
-package com.example.shell.fx.layouts.files.view.comboBoxes;
-
-import javafx.scene.control.ComboBox;
-
-import java.util.List;
-
-public abstract class ComboBoxTemplate {
-    // Шаблонний метод
-    public ComboBox<String> createMenu(List<String> menuItems) {
-        ComboBox<String> menu = createBaseMenu(menuItems);
-        customizeMenu(menu);
-        return menu;
-    }
-
-    // Абстрактний метод, який потрібно реалізувати в підкласах
-    protected abstract ComboBox<String> createBaseMenu(List<String> menuItems);
-
-    // Абстрактний метод для налаштування меню
-    protected abstract void customizeMenu(ComboBox<String> menu);
-}
