--- src/main/java/com/example/shell/fx/layouts/files/view/views/FilesViewFactory.java
@@ -1,11 +1,0 @@
-package com.example.shell.fx.layouts.files.view.views;
-
-import com.example.shell.fx.layouts.files.view.views.Impl.FilesView;
-
-public abstract class FilesViewFactory {
-    public FilesView create() {
-        return createFilesView();
-    }
-
-    protected abstract FilesView createFilesView();
-}
