--- src/main/java/com/example/shell/fx/layouts/files/view/views/Impl/FilesView.java
@@ -1,14 +1,0 @@
-package com.example.shell.fx.layouts.files.view.views.Impl;
-
-import javafx.scene.Node;
-
-import java.io.File;
-import java.util.List;
-
-public interface FilesView {
-    Node getNode();
-
-    void addFiles(List<File> files);
-
-    String getName();
-}
