--- src/main/java/com/example/shell/fx/layouts/files/view/views/Impl/FilesViewType.java
@@ -1,6 +1,0 @@
-package com.example.shell.fx.layouts.files.view.views.Impl;
-
-public enum FilesViewType {
-    LIST,
-    TABLE
-}
