+++ src/main/java/com/example/shell/fx/layouts/files/views/Impl/FilesView.java
@@ -1,0 +1,14 @@
+package com.example.shell.fx.layouts.files.views.Impl;
+
+import javafx.scene.Node;
+
+import java.io.File;
+import java.util.List;
+
+public interface FilesView {
+    Node getNode();
+
+    void addFiles(List<File> files);
+
+    String getName();
+}
