+++ src/main/java/com/example/shell/fx/layouts/files/views/Impl/FilesViewType.java
@@ -1,0 +1,6 @@
+package com.example.shell.fx.layouts.files.views.Impl;
+
+public enum FilesViewType {
+    LIST,
+    TABLE
+}
